requirements.txt
certifi==2019.6.16
chardet==3.0.4
facebook-sdk==3.1.0
ffmpeg==1.4
ffmpeg-python==0.2.0
future==0.17.1
idna==2.8
pylivestream==1.9.4
requests==2.22.0
urllib3==1.25.3
